fix: migrate storage from browser version
When bitte-wallet was a browser wallet, it was storing the information
on the logged-in account in localStorage under a specific key. When we
migrated it to be an injected wallet, we stored the account information
in a different storage key.

If an application migrates from wallet selector v8 to v9, users that
are already logged-in end up in a weird state, in which the wallet
selector detects the user was readily logged in, but cannot find
the username.

This commit transforms the old storage key into the new one, so the
wallet selector can correctly detect which user is logged in.